requirements.txt
gspread==5.0.0
google-auth==2.3.3
# Include any other libraries you may need for your script
